anomalydetector/requirements.txt
quixstreams==3.2.1
python-dotenv
scikit-learn